Исправления: Перенесено в отдельный класс
