+++ app.rb
@@ -1,87 +1,37 @@
 # frozen_string_literal: true
+
+require_relative 'time_format'
 class App
-  attr_accessor :status, :body, :header
 
   def call(env)
-    request_parser(env)
-    [@status, @header, @body]
-  end
-
-  private
-
-  def request_parser(env)
     req = Rack::Request.new(env)
     path = req.path
 
+    pp "path = #{path}"
     case path
     when '/'
       root
     when '/time'
-      server_time(req)
+      time_formatter(req)
     else
       redirect_to_root
     end
   end
 
-  def server_time(req)
-    @body = [Time.now.to_s]
-    @status = 200
-    @header = headers
-    return if req.params.empty?
+  private
 
-    formatter(req.params['format'])
+  def time_formatter(req)
+    tf = TimeFormat.new(req)
+    tf.call
+    response(tf.status, tf.header, tf.body)
   end
 
-  def redirect_to_root
-    @status = 302
-    @header = { 'Location' => 'http://localhost:9292' }
-    @body = []
-  end
-
-  def formatter(format)
-    formatter = ''
-    unknown_formats = ''
-    return if format.nil? || format.empty?
-
-    query_params = format.split(',')
-    query_params.each do |qp|
-      formatter += case qp
-                   when 'year'
-                     ' %Y '
-                   when 'month'
-                     ' %m '
-                   when 'day'
-                     ' %d '
-                   when 'hour'
-                     ' %H '
-                   when 'minute'
-                     ' %m '
-                   when 'seconds'
-                     ' %S '
-                   else
-                     unknown_formats += unknown_formats.empty? ? qp : ", #{qp}"
-                     ''
-                  end
-    end
-    unknown_formats = "Unknown format: #{unknown_formats}" unless unknown_formats.empty?
-    return unknown_format(unknown_formats) unless unknown_formats.empty?
-
-    @body = [Time.now.strftime(formatter)]
-  end
-
-  def unknown_format(unknown)
-    @status = 400
-    @body = [unknown]
+  def response(status, header, body)
+    Rack::Response.new(body, status, header).finish
   end
 
   def perform_request
     sleep rand(2..3)
-  end
-
-  def root
-    @status = 200
-    @header = { 'Content-Type' => 'text/html' }
-    @body = ['Time server', '<br>', '<a href="/time?format=year,month,day">Time?</a>']
   end
 
   def status
@@ -95,4 +45,17 @@ class App
   def body
     ["Welcome aboard!\n"]
   end
+
+  def redirect_to_root
+    @status = 302
+    @header = { 'Location' => 'http://localhost:9292' }
+    @body = []
+    response(@status, @header, @body)
+  end
+
+  def root
+    @body = ['Time server', '<br>', '<a href="/time?format=year,month,day">Time?</a>']
+    response(200, { 'Content-Type' => 'text/html' }, @body)
+  end
+
 end
